Upgrade JDom to version 2.0.6
JTS is using a version of JDom that has long been EOL, having been released in Feb 2012.

This upgrades JDom to version 2.0.6, as part of this upgrade I've replaced LineNumberSAXBuilder with the regular jdom SAXBuilder configured with a LocatedJDOMFactory. This is neccesary because LineNumberSAXBuilder overrides the element methods that are now final in the class being extend.

Signed-off-by: Mark Prins <mprins@users.sf.net>