RFC: Tests for ISO compliance
This PR is meant to start a discussion about where JTS is and isn't
confirming to the ISO spec.  These test cases are meant to explore the
"basic" scenarios thoroughly.  They aren't meant to test numerical
stability or cpmplex cases, but are meant to test conditions like
nulls and NaNs.

The tests were designed to pass in JTS.  Where I've noticed a
departure from the ISO spec, or unexpected behavior, I've put an
"XXX" comment.