GeometrySnapper calculate tolerance from largest part of geometry collections
Currently the snap tolerance has been calculated from the geometry envelope.
But this can result in a large tolerance for geometry collections where the
envelope is large but that contain parts are much smaller.

In the geometry collection case we can calculate the tolerance that ought to
be applied to each part, and use the maximum of those values.

Signed-off-by: Glenn Walbran <glennw@catalyst.net.nz>